Lab-02/app/requirements.txt
flask==3.0.3
gunicorn==22.0.0
azure-storage-blob==12.19.0